requirements.txt
redis==5.0.4
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
inflect==7.0.0
orjson==3.9.15
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
sniffio==1.3.1
sqlacodegen==3.0.0rc3
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
requests==2.31.0
httpx==0.27.0
websocket-client==1.8.0
schedule==1.2.1
loguru==0.7.2
apscheduler==3.10.4



